pyproject.toml
[project]
name = "bot1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dateparser>=1.2.1",
    "google-genai>=1.20.0",
    "instaloader>=4.14.1",
    "pika>=1.3.2",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.3",
    "the-retry>=0.1.1",
]
